src/main.py
import logging

from dotenv import load_dotenv

from adapters.api import app


load_dotenv()
logging.basicConfig(
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    level=logging.DEBUG)
