requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
Inject==3.5.4
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.3
Werkzeug==0.16.0
